.history/requirements_20211017185843.txt
Django==3.2.8
djangorestframework==3.12.4
django-rest-swagger==2.2.0
numpy==1.21.2
scikit-learn==1.0
joblib==1.1.0